Codefactor whining fix
